% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{filter_duplicates}
\alias{filter_duplicates}
\alias{filter_unique}
\title{Filter duplicates on desired variables}
\usage{
filter_duplicates(data, ...)

filter_unique(data, ...)
}
\arguments{
\item{data}{a tbl}

\item{...}{Variables to find duplicates of}
}
\description{
Filter duplicates or unique values. These are
wrappers around a `group_by` then `filter`.
}
\examples{

df <- data.frame(grp1 = c(1, 1, 2, 2, 3), grp2 = c(1, 2, 1, 2, 2))

df \%>\% filter_duplicates(grp1)
df \%>\% filter_duplicates(grp2)

df \%>\% filter_unique(grp1, grp2)
}
