% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_multlivalue.R
\name{gather_multivalue}
\alias{gather_multivalue}
\title{tidyr's gather for multiple values}
\usage{
gather_multivalue(data, vars, key = "key", regex = "^([a-zA-Z]+)(\\\\d+)$")
}
\arguments{
\item{data}{a data frame}

\item{vars}{A selection of columns. If empty, all variables are selected.
You can supply bare variable names, select all variables between x and z with x:z,
exclude y with -y. For more options, see the dplyr::select() documentation.}

\item{key}{key}

\item{regex}{a regular expression used to extract the desired values}
}
\description{
Extension of `tidyr::gather` when there is a key with multiple values. Same
purpose as Stata's `reshape` long.
}
\examples{

scores <- data.frame(
id = LETTERS[1:5],
age2000 = 11:15,
age2010 = 21:25,
scores2000 = 96:100,
scores2010 = 100:96)

gather_multivalue(scores, -id)
gather_multivalue(scores, age2000:scores2010)


scores2 <- data.frame(
id = LETTERS[1:5],
age_2000 = 11:15,
age_2010 = 21:25,
scores_2000 = 96:100,
scores_2010 = 100:96)

gather_multivalue(scores2, -id, regex = "([a-z]+)_(\\\\d+)")
}
