% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread_multivalue.R
\name{spread_multivalue}
\alias{spread_multivalue}
\title{tidyr's spread for multiple values}
\usage{
spread_multivalue(data, key, values, sep = "")
}
\arguments{
\item{data}{a data frame}

\item{key}{key, a string}

\item{values}{A selection of columns. If empty, all variables are selected.
You can supply bare variable names, select all variables between x and z with x:z,
exclude y with -y. For more options, see the dplyr::select() documentation.}

\item{sep}{separator to use between key and values}
}
\description{
Extension of `tidyr::spread` when there is a key with multiple values. Same
purpose as Stata's `reshape` wide.
}
\examples{

scores <- data.frame(
id = LETTERS[1:5],
age2000 = 11:15,
age2010 = 21:25,
scores2000 = 96:100,
scores2010 = 100:96)

# gather_multivalue(scores, "year", -id)
scores_long <- gather_multivalue(scores, "year", age2000:scores2010)

spread_multivalue(scores_long, "year", c(age, scores))
}
